Revert "Dev to master (#54)"
This reverts commit 42455efe7924854305e0a55fd506c95139fe006d.
