+++ src/components/FormListModal.vue
@@ -1,8 +1,8 @@
 <template>
         <div class="cftf-form-conatiner">
-          <b-button class="cft-form-button" v-b-modal.modal-multi-1>
+          <button type="button" class="btn btn-light cft-form-title" v-b-modal.modal-multi-1>
               <h4 ref="btn-show"> <i class="fa fa-wpforms"></i> Forms</h4>
-          </b-button>
+          </button>
             <b-modal
               ref="modal-1"
               id="modal-multi-1"
