+++ src/components/tasklist.vue
@@ -1,19 +1,143 @@
 <template> 
 <b-container fluid class="task-outer-container">
-  <Header
-  v-if="token  && bpmApiUrl"
-  :token="token"
-  :bpmApiUrl="bpmApiUrl"
-  />
-    <b-row class="cft-service-task-list mt-1">
-      <b-col xl="3" lg="3" md="12" class="cft-first">
-        <LeftSider
-          v-if="token  && bpmApiUrl"
-          :token="token"
-          :formsflowaiApiUrl="formsflowaiApiUrl"
-          :formIOApiUrl="formIOApiUrl"
-          :bpmApiUrl="bpmApiUrl"
-        />
+  <div class="main-filters my-2 mb-1">
+    <div 
+      class="cft-filter-dropdown mx-2"
+    >
+        <button class="cft-filter-dropbtn mr-0" @click="toggleshowfilter">
+          <i class="bi bi-filter-square"/>
+        </button>
+        <div v-if="showfilter" class="cft-filter-dropdown-content">
+          <b-list-group  v-if="filterList.length">
+            <b-list-group-item v-for="(filter, idx) in filterList" :key="filter.id"
+            @click="togglefilter(filter, idx)"
+            :class="{'cft-filter-selected': idx == activefilter}">
+              {{filter.name}}
+            </b-list-group-item>
+          </b-list-group>
+          <div v-else>
+            <i class="bi bi-exclamation-circle-fill"></i>
+              No Filters found  
+          </div>
+        </div>
+      </div>
+
+      <FormListModal :token="token" :bpmApiUrl="bpmApiUrl"/>          
+
+      <div class="cft-first">
+        <!-- Sorting section -->
+				<div id="cftf-dpdown-container" class="mx-2">
+					<div class="cftf-dpdown-box mr-2" v-for="(sort, idx) in sortList" :key="sort.sortBy">
+              <span v-if="sortList.length>1"
+                class="cftf-exit-button"
+                title="Remove Sorting" 
+                @click="deleteSort(sort, index)"
+              >
+                  <i class="bi bi-x"></i>
+              </span>
+							<span
+               class="cftf-span-element"
+                @click="showUpdateSortOptions(idx)"
+              >{{sortList[idx]["label"]}}</span>
+              <div
+                v-if="showSortListDropdown[idx]"
+                class="cft-sort-items"
+              >
+                <div v-for="s in sortOptions" :key="s.sortBy"
+                 @click="updateSort(s,idx)"
+                 class="mb-2 cft-sort-element"
+                >
+                  {{s.label}}
+                </div>
+              </div>
+                <a v-if="sort.sortOrder==='asc'" @click="toggleSort(idx)" href="#" title="Ascending">
+                  <i class="bi bi-chevron-up cftf-arrow"></i>
+                </a>
+                <a v-else @click="toggleSort(idx)"  href="#" title="Descending">
+                  <i class="bi bi-chevron-down cftf-arrow"></i>
+                </a>
+            </div>
+          <i v-if="updateSortOptions.length===0"
+           class="fa fa-plus fa-sm click-element cftf-add-sorting"
+           @click="showaddSortListOptions"
+           title="Add sorting"></i>
+          <TaskSortOptions
+           :sortOptions="sortOptions"
+           :showSortListDropdown="showaddNewSortListDropdown"
+           @add-sort="addSort"
+          >
+          </TaskSortOptions>
+          </div>
+				</div>					 
+  </div>
+	<b-row class="cft-service-task-list mt-1">
+		<b-col xl="3" lg="3" md="12" class="cft-first">
+      <TaskListSearch
+        @call-search-api="callSearchApi"
+        @call-search-date-api="callSearchDateApi"
+        @call-process-variables-api="callProcessVariablesApi"
+        @call-task-variables-api="callTaskVariablesApi"
+        :tasklength="tasklength"
+      />
+        <!-- Task list section -->
+        <b-list-group class="cft-list-container" v-if="tasks && tasks.length">
+          <b-list-group-item
+            button
+            v-for="(task, idx) in tasks"
+            v-bind:key="task.id"
+            v-on:click="toggle(idx)"
+            :class="{ 'cft-selected': idx == activeIndex }"
+          >
+            <div @click="setselectedTask(task.id)" class="cft-select-task">
+              <h5 class="cft-task-title" data-title='Task Name'>{{ task.name }}</h5>
+
+              <div class="cft-task-details-assign assigne-details ">
+                <div class='cft-process-title' data-tile='Process Definition Name'>
+                  {{
+                    getProcessDataFromList(
+                      getProcessDefinitions,
+                      task.processDefinitionId,
+                      "name"
+                    )
+                  }}
+                </div>
+                <div class="cft-task-assignee" data-title="Task assignee">
+                  {{ task.assignee }}
+                </div>
+              </div>
+              <div class="cft-task-details-assign font-11">
+                <div class="created-details">
+                  <div>
+                    <span class="cft-due-date" :data-title="getExactDate(task.due)" v-if="task.due">
+                      Due {{ timedifference(task.due) }},
+                    </span>
+                    <span class="cft-due-date" :data-title="getExactDate(task.followUp)" v-if="task.followUp">
+                      Follow-up {{ timedifference(task.followUp) }},
+                    </span>
+                    <span class="cft-due-date" :data-title="getExactDate(task.created)" v-if="task.created">
+                      Created {{ timedifference(task.created) }}
+                    </span>
+                  </div>
+                  <div class="cft-priority" data-title="priority">
+                    {{ task.priority }}
+                  </div>
+                </div>
+              </div>
+            </div>
+          </b-list-group-item>
+          <b-pagination-nav
+            :link-gen="linkGen"
+            :number-of-pages="numPages"
+            v-model="currentPage"
+            class="cft-paginate"
+          />
+        </b-list-group>
+        <b-list-group cols="3" v-else>
+          <b-row class="cft-not-selected mt-2 ml-1 row">
+            <i class="bi bi-exclamation-circle-fill" scale="1"></i>
+            <p>No tasks found in the list.</p>
+          </b-row>
+        </b-list-group>
       </b-col>
       <!-- Task Detail section -->
       <b-col v-if="selectedTaskId" lg="9" md="12">
@@ -70,66 +194,25 @@
                     class="cft-groups"
                     data-title="groups"
                 >
-                <i class="bi bi-grid-3x3-gap-fill"></i>
+                    <i class="bi bi-grid-3x3-gap-fill"></i>
                     {{ String(groupListNames) }}
                 </div>				 
                 <div
-                  v-b-modal.AddGroupModal
-                  class="cft-groups"
-                  data-title="groups"
-                  v-else
+                    v-b-modal.AddGroupModal
+                    class="cft-groups"
+                    data-title="groups"
+                    v-else
                 >
-                  <i class="bi bi-grid-3x3-gap-fill"></i> Add Groups
+                <i class="bi bi-grid-3x3-gap-fill"></i> Add Groups
                 </div>
-                <b-modal
-                  id="AddGroupModal"
-                  ref="modal"
-                  title="Manage Groups"
-                  ok-title="Close"
-                  ok-only
-                >
-                  <div class="modal-text">
-                    <i class="bi bi-exclamation-circle"></i>
-                    You can add a group by typing a group ID into the input
-                    field and afterwards clicking the button with the plus sign.
-                    <b-row class="mt-3 mb-3">
-                      <b-col>
-                        <b-button
-                          variant="primary"
-                          @click="addGroup"
-                          :disabled='!setGroup'
-                        >
-                          <span>Add a group</span>
-                          <span>
-                            <i class="bi bi-plus"></i>
-                          </span>
-                        </b-button>
-                      </b-col>
-                      <b-col>
-                        <b-form-input
-                          type="text"
-                          placeholder="Group ID"
-                          v-model="setGroup"
-                          v-on:keyup.enter="addGroup"
-                        >
-			</b-form-input>
-                      </b-col>
-                    </b-row>
-                    <b-row>
-                      <b-col v-if="groupList.length">
-                        <ul v-for="g in groupList" :key="g.groupId">
-                          <div class="mt-1">
-                            <i
-                              class="fa fa-times mr-2 click-element"
-                              @click="deleteGroup(g.groupId)"
-                            ></i>
-                            <span>{{ g.groupId }}</span>
-                          </div>
-                        </ul>
-                      </b-col>
-                    </b-row>
-                  </div>
-                </b-modal>
+                <TaskListGroup
+                  :token='token'
+                  :bpmApiUrl='bpmApiUrl'
+                  :task="task"
+                  :groupList='groupList'
+                  @add-group='addGroup'
+                  @delete-group='deleteGroup'
+                />             
               </b-col>
               <b-col>
                 <div
@@ -189,16 +272,15 @@
                     </b-overlay>
                   </div>
                 </b-tab>
-                <b-tab title="History">
-                  <TaskHistory :taskHistoryList='taskHistoryList' :applicationId="applicationId"/>
-                </b-tab>
+                <b-tab title="History"></b-tab>
                 <!-- Process diagram -->
                 <b-tab
                   class="cft-diagram-container"
                   id="diagramContainer"
                   title="Diagram"
+                  @click="fetchxmldiagram"
                 >
-                  <div class="height-100 cft-canvas-container diagram-full-screen" id="canvas"></div>
+                  <div class="height-100 cft-canvas-container" id="canvas"></div>
                 </b-tab>
               </b-tabs>
             </div>
@@ -220,24 +302,30 @@
 </template>
 
 <script lang="ts">
+
+interface Payload{
+    active: boolean;
+    sorting: Array<object>;
+    orQueries?: Array<object>;
+    maxResults?: number;
+    firstResult?: number;
+  }
+
 import 'bootstrap-icons/font/bootstrap-icons.css';
-import 'font-awesome/scss/font-awesome.scss';
+import "font-awesome/scss/font-awesome.scss";
 import 'formiojs/dist/formio.full.min.css'
 import 'vue2-datepicker/index.css';
 import 'semantic-ui-css/semantic.min.css';
 import '../styles/user-styles.css'
 import '../styles/camundaFormIOTasklist.scss'
 import { Component, Prop, Vue, Watch } from 'vue-property-decorator'
-// eslint-disable-next-line sort-imports
-import Header from './layout/header.vue'
-import LeftSider from './layout/left-sider.vue'
-
 import {
   TASK_FILTER_LIST_DEFAULT_PARAM,
   decodeTokenValues,
   findFilterKeyOfAllTask,
   getFormattedDateAndTime,
   getTaskFromList,
+  searchQuery,
   sortingList,
 } from '../services/utils';
 import BpmnJS from 'bpmn-js';
@@ -246,17 +334,15 @@ import DatePicker from 'vue2-datepicker'
 import { Form } from 'vue-formio';
 import FormListModal from './FormListModal.vue';
 import Modeler from 'bpmn-js/lib/Modeler';
-import {Payload} from '../services/TasklistTypes';
-import SocketIOService from '../services/SocketIOServices';
-import TaskHistory from '../components/TaskHistory.vue';
-import TaskListSearch from '../components/TaskListSearch.vue';
-import {authenticateFormio} from '../services/formio-token';
+import SocketIOService from "../services/SocketIOServices";
+import TaskListGroup from "../components/TasklistGroup.vue";
+import TaskListSearch from "../components/TasklistSearch.vue";
+import TaskSortOptions from '../components/TasklistSortoptions.vue';
+import {authenticateFormio} from "../services/formio-token";
 import {getFormDetails} from '../services/get-formio';
 import {getISODateTime} from '../services/format-time';
-import {getformHistoryApi} from '../services/formsflowai-api';
-import moment from 'moment';
-import {searchQuery} from '../services/search-constants';
-import vueBpmn from 'vue-bpmn';
+import moment from "moment";
+import vueBpmn from "vue-bpmn";
 
 
 @Component({
@@ -264,13 +350,12 @@ import vueBpmn from 'vue-bpmn';
     formio: Form,
     DatePicker,
     FormListModal,
-    TaskHistory,
+    TaskListGroup,
     TaskListSearch,
+    TaskSortOptions,
     vueBpmn,
     Modeler,
     BpmnJS,
-    Header,
-    LeftSider
   },
 })
 export default class Tasklist extends Vue {
@@ -284,7 +369,7 @@ export default class Tasklist extends Vue {
   @Prop() private formsflowaiUrl!: string;
   @Prop() private formIOUserRoles!: string;
   @Prop() private userName!: string;
-  @Prop({default:'formflowai'}) private webSocketEncryptkey !: string;
+  @Prop({default:'formflowai'}) private WEBSOCKET_ENCRYPT_KEY !: string;
 
   private tasks: Array<object> = [];
   private fulltasks: Array<object> = [];
@@ -323,7 +408,7 @@ export default class Tasklist extends Vue {
   private groupList = [];
   private groupListNames: Array<string> | null = null;
   private groupListItems: string[] = [];
-  private userEmail = 'external';
+  private userEmail = "external";
   private selectedfilterId = '';
   private xmlData!: string;
   private sortList = TASK_FILTER_LIST_DEFAULT_PARAM;
@@ -338,7 +423,6 @@ export default class Tasklist extends Vue {
     sorting: TASK_FILTER_LIST_DEFAULT_PARAM,
   };
   private showUserList = false;
-  private taskHistoryList: Array<object> = [];
   
 @Watch('token')
   ontokenChange (newVal: string) {
@@ -369,17 +453,19 @@ checkPropsIsPassedAndSetValue() {
   if (!this.formsflowaiUrl || this.formsflowaiUrl === "") {
     console.warn("formsflow.ai URL prop not passed");
   }
-  if(!this.webSocketEncryptkey || this.webSocketEncryptkey === ""){
+  if(!this.WEBSOCKET_ENCRYPT_KEY || this.WEBSOCKET_ENCRYPT_KEY === ""){
     console.warn('WEBSOCKET_ENCRYPT_KEY prop not passed')
   }
   const engine = "/engine-rest";
   const socketUrl = "/forms-flow-bpm-socket";
   localStorage.setItem("bpmApiUrl", this.bpmApiUrl + engine);
   localStorage.setItem("authToken", this.token);
-  const currentUrl = window.location.protocol + '//' +  window.location.host
-  localStorage.setItem("formsflow.ai.url", currentUrl);
+  localStorage.setItem("formsflow.ai.url", this.formsflowaiUrl);
   localStorage.setItem("formsflow.ai.api.url", this.formsflowaiApiUrl);
   localStorage.setItem("formIOApiUrl", this.formIOApiUrl);
+  localStorage.setItem("bpmSocketUrl", this.bpmApiUrl + socketUrl)
+  localStorage.setItem("websocketEncryptkey", this.WEBSOCKET_ENCRYPT_KEY)
+
   const val = decodeTokenValues(
     this.token,
     this.userName,
@@ -394,13 +480,61 @@ timedifference(date: Date) {
   return moment(date).fromNow();											
 }													   
 
+getProcessDataFromList(processList: any[], processId: string, dataKey: string) {
+  const process = processList.find((process) => process.id === processId);
+  return process && process[dataKey];
+}
+
+setselectedTask(task: string) {
+  this.selectedTaskId = task;
+  this.fetchData();
+}
+getExactDate(date: Date) {
+  return getFormattedDateAndTime(date);
+}
 toggle(index: number) {
   this.activeIndex = index;						  
+}
+
+toggleshowfilter() {
+  this.showfilter = ! this.showfilter;
+}
+
+togglefilter(filter: any, index: number) {
+  this.activefilter = index;
+  this.fetchTaskList(filter.id, this.payload);
+  this.showfilter = false;
 }
 
 toggleassignee()  {
   this.editAssignee = ! this.editAssignee;
   this.userSelected = this.task.assignee;
+}
+
+cftShowUserList() {
+  this.showUserList = !this.showUserList;
+}
+
+callSearchApi(item: any) {
+  this.payload["orQueries"] = item;
+  this.fetchTaskList(this.selectedfilterId, this.payload);
+}
+
+callSearchDateApi(item: any) {
+  this.payload["orQueries"] = item;
+  this.fetchTaskList(this.selectedfilterId, this.payload);
+}
+
+callTaskVariablesApi(item: any) {
+  searchQuery["taskVariables"].push(item);
+  this.payload["orQueries"] = [searchQuery];
+  this.fetchTaskList(this.selectedfilterId, this.payload);
+}
+
+callProcessVariablesApi(item: any) {
+  searchQuery["processVariables"].push(item);
+  this.payload["orQueries"] = [searchQuery];
+  this.fetchTaskList(this.selectedfilterId, this.payload);
 }
 
 onFormSubmitCallback() {
@@ -411,17 +545,10 @@ onFormSubmitCallback() {
 }
 
 addGroup() {
-  CamundaRest.createTaskGroupByID(
-    this.token,
-    this.task.id,
-    this.bpmApiUrl,
-    { userId: null, groupId: this.setGroup, type: "candidate" }
-  ).then(() => {
-    this.getGroupDetails();
-    this.reloadCurrentTask();
-    this.setGroup = null;
-  });
+  this.getGroupDetails();
+  this.reloadCurrentTask();
 }
+
 getGroupDetails() {
   CamundaRest.getTaskGroupByID(this.token, this.task.id, this.bpmApiUrl).then(
     (response) => {
@@ -437,14 +564,10 @@ getGroupDetails() {
     }
   );
 }
-deleteGroup(groupid: string) {		 
-  CamundaRest.deleteTaskGroupByID(this.token, this.task.id, this.bpmApiUrl, {
-    groupId: groupid,
-    type: "candidate",
-  }).then(() => {
-    this.getGroupDetails();
-    this.reloadCurrentTask();
-  });
+
+deleteGroup() {
+  this.getGroupDetails();
+  this.reloadCurrentTask();
 }
 
 onBPMTaskFormSubmit(taskId: string) {
@@ -509,14 +632,17 @@ getBPMTaskDetail(taskId: string) {
   };
 
   reloadTasks() {
+    //used to unSelect the task and refresh taskList
     this.selectedTaskId = "";
     this.fetchTaskList(this.selectedfilterId, this.payload);
   }
 
   reloadCurrentTask() {
+    //used to refresh selected task and taskList
     this.getBPMTaskDetail(this.task.id);
     this.fetchTaskList(this.selectedfilterId, this.payload);
   }
+ 
 
   onClaim() {
     CamundaRest.claim(
@@ -575,6 +701,39 @@ getBPMTaskDetail(taskId: string) {
     });
   }
 
+  fetchInittasklist(filterId: string, payload: object){
+    CamundaRest.filterTaskList(
+      this.token,
+      filterId,
+      payload,
+      this.bpmApiUrl
+    ).then((result) => {
+      this.tasks = result.data;
+      this.fulltasks = result.data;
+      this.numPages =  Math.ceil(result.data.length / this.perPage);
+    })
+  }
+
+  // fetchPaginationTaskElements(){
+  //   this.tasks = this.fulltasks.slice(
+  //     (this.currentPage - 1) * this.perPage,
+  //     this.currentPage * this.perPage
+  //   );
+  // }
+
+  numberOfPages() {
+    if (Math.ceil(this.tasks.length / this.perPage) > 1)
+      return Math.ceil(this.tasks.length / this.perPage);
+    else {
+      return 15;
+    }
+  }
+
+  linkGen() {
+    this.fetchTaskList(this.selectedfilterId, this.payload);
+  }
+ 
+
   getOptions(options: any) {
     const optionsArray: {
       sortOrder: string;
@@ -591,6 +750,61 @@ getBPMTaskDetail(taskId: string) {
       }
     });
     return optionsArray;
+  }
+
+  addSort(sort: any) {
+    this.sortList.push(sort);
+    if (this.sortList.length === sortingList.length) {
+      this.updateSortOptions = this.sortOptions;
+    } else {
+      this.sortOptions = this.getOptions(this.sortList);
+    }
+    this.showaddNewSortListDropdown = false;									  
+  }
+
+  showaddSortListOptions() {
+    this.showaddNewSortListDropdown = !this.showaddNewSortListDropdown;
+    this.sortOptions = this.getOptions(this.sortList);
+  }
+
+  showUpdateSortOptions(index: number) {
+    for(let i =0; i<6;i++){
+      if(this.showSortListDropdown[i]===true){
+        this.showSortListDropdown[i] = false;
+      }
+    }
+    this.showSortListDropdown[index] = !this.showSortListDropdown[index];
+    this.sortOptions = this.getOptions(this.sortList);
+    this.setupdateSortListDropdownindex = index;
+  }
+
+  updateSort(sort: any, index: number) {
+    this.sortList[index].label = sort.label;
+    this.sortList[index].sortBy = sort.sortBy;
+
+    this.sortOptions = this.getOptions(this.sortList);
+    this.showSortListDropdown[index] = false;
+    this.payload["sorting"] = this.sortList;
+    this.fetchTaskList(this.selectedfilterId, this.payload);
+  }
+
+  deleteSort(sort: any, index: number) {
+    this.sortList.splice(index, 1);
+    this.updateSortOptions = [];
+    this.sortOptions = this.getOptions(this.sortList);
+    this.payload["sorting"] = this.sortList;
+    this.fetchTaskList(this.selectedfilterId, this.payload);
+  }
+
+  toggleSort(index: number) {
+    if (this.sortList[index].sortOrder === "asc")
+      this.sortList[index].sortOrder = "desc";
+  
+    else {
+      this.sortList[index].sortOrder = "asc";
+    }
+    this.payload["sorting"] = this.sortList;
+    this.fetchTaskList(this.selectedfilterId, this.payload);
   }
 
   updateFollowUpDate() {
@@ -634,6 +848,9 @@ getBPMTaskDetail(taskId: string) {
           console.error("Error", error);
         });
     }
+    else {
+      console.warn("Due date error");
+    }
   }
 
   removeDueDate() {
@@ -662,6 +879,19 @@ getBPMTaskDetail(taskId: string) {
     })
   }
 
+  fetchxmldiagram() {
+    CamundaRest.getProcessDiagramXML(
+      this.token,
+      this.processDefinitionId,
+      this.bpmApiUrl
+    ).then(async (res) => {
+      this.xmlData = res.data.bpmn20Xml;
+      const modeler = new Modeler({ container: "#canvas" });
+      await modeler.importXML(this.xmlData);
+    });
+
+  }
+
   fetchData() {
     if (this.selectedTaskId) {
       this.task = getTaskFromList(this.tasks, this.selectedTaskId);
@@ -671,28 +901,13 @@ getBPMTaskDetail(taskId: string) {
         this.selectedTaskId,
         this.bpmApiUrl
       ).then((result) => {
+        this.processDefinitionId = result.data.processDefinitionId;
         CamundaRest.getProcessDefinitionById(
           this.token,
           result.data.processDefinitionId,
           this.bpmApiUrl
         ).then((res) => {
           this.taskProcess = res.data.name;
-        });
-
-        CamundaRest.getVariablesByProcessId(
-          this.token,
-          result.data.processInstanceId,
-          this.bpmApiUrl
-        )
-
-        CamundaRest.getProcessDiagramXML(
-          this.token,
-          result.data.processDefinitionId,
-          this.bpmApiUrl
-        ).then(async (res) => {
-          this.xmlData = res.data.bpmn20Xml;
-          const modeler = new Modeler({ container: "#canvas" });
-          await modeler.importXML(this.xmlData);
         });
       });
 
@@ -702,15 +917,7 @@ getBPMTaskDetail(taskId: string) {
         this.selectedTaskId,
         this.bpmApiUrl
       ).then((result) => {
-        if(result.data && result.data["applicationId"].value) {
-          getformHistoryApi(this.formsflowaiApiUrl, result.data["applicationId"].value, this.token)
-            .then((r)=> {
-              this.taskHistoryList = r.data.applications;
-            })
-        }
-        else {
-          console.warn("The selected task has no applicationid")
-        }
+        // handle case when applicationid and formUrl null
         this.applicationId = result.data["applicationId"].value;
         this.formioUrl = result.data["formUrl"].value;
         const { formioUrl, formId, submissionId } = getFormDetails(
@@ -721,17 +928,12 @@ getBPMTaskDetail(taskId: string) {
         this.submissionId = submissionId;
         this.formId = formId;
         this.showfrom = true;
-        this.userSelected = this.task.assignee;
       });
+      this.userSelected = this.task.assignee;									   
     }
   }
   
   mounted() {
-    this.$root.$on('call-fetchData', (para: any) => {
-      this.selectedTaskId = para.selectedTaskId
-      this.fetchData()
-    })
-
     this.checkPropsIsPassedAndSetValue();
     authenticateFormio(
       this.formIOResourceId,
@@ -744,17 +946,16 @@ getBPMTaskDetail(taskId: string) {
       this.filterList = response.data;
       this.selectedfilterId = findFilterKeyOfAllTask(this.filterList, "name", "All tasks");
       this.fetchTaskList(this.selectedfilterId, this.payload);
-      this.fetchData();
     });
 
+    this.fetchData();
     if(SocketIOService.isConnected()) {
       SocketIOService.disconnect();
     }
-    SocketIOService.connect(this.webSocketEncryptkey, (refreshedTaskId: any)=> {
+    SocketIOService.connect((refreshedTaskId: any)=> {
       if(this.selectedfilterId){
         //Refreshes the Task
         this.fetchTaskList(this.selectedfilterId, this.payload);
-        this.fetchData();
       }
       if(this.selectedTaskId && refreshedTaskId===this.selectedTaskId){
         this.fetchData()
