+++ src/services/SocketIOServices.ts
@@ -3,9 +3,7 @@ import Stomp from 'stompjs'
 import AES from 'crypto-js/aes';
 
 let stompClient: any = null;
-const engine = "/engine-rest";
-const socketUrl = "/forms-flow-bpm-socket";
-const BPM_BASE_URL_SOCKET_IO = localStorage.getItem('bpmApiUrl') ? localStorage.getItem('bpmApiUrl')?.replace(engine, socketUrl) : ''
+const BPM_BASE_URL_SOCKET_IO = localStorage.getItem('bpmSocketUrl');
 const token: any = localStorage.getItem('authToken');
 const WEBSOCKET_ENCRYPT_KEY: any = localStorage.getItem('websocketEncryptkey');
 
@@ -14,8 +12,8 @@ const isConnected = ()=>{
   return stompClient?.connected||null;
 };
 
-const connect = (encryptKey: any, reloadCallback: any)=>{
-  const accessToken= AES.encrypt(token, encryptKey).toString();
+const connect = (reloadCallback: any)=>{
+  const accessToken= AES.encrypt(token, WEBSOCKET_ENCRYPT_KEY).toString();
   const socketUrl=`${BPM_BASE_URL_SOCKET_IO}?accesstoken=${accessToken}`;
   const socket = new SockJS(socketUrl);
   stompClient = Stomp.over(socket);
