--- src/services/TasklistTypes.ts
@@ -1,7 +1,0 @@
-export interface Payload{
-    active: boolean;
-    sorting: Array<object>;
-    orQueries?: Array<object>;
-    maxResults?: number;
-    firstResult?: number;
-  }
