+++ src/styles/camundaFormIOFormList.scss
@@ -2,12 +2,6 @@
   background-color: white;
   text-align: left;
   font-size: 16px;
-
-  .cft-form-button {
-    background-color: #1a5a96;
-    border: #38598a !important;
-    color: white !important;
-  }
   
   .bg-default {
     background-color: #38598a !important;
