+++ src/styles/camundaFormIOTasklist.scss
@@ -437,8 +437,5 @@
   border-color: none !important;
   box-shadow: none;
 }
-.diagram-full-screen {
-  height: 730px;
-}
 
 }
